lambda/get_daily_metrics/requirements.txt
requests==2.31.0
mysql-connector-python==8.0.32
web3
